+++ src/types/TSearch/TRecentSearchListTypes.ts
@@ -1,0 +1,6 @@
+import { RefObject } from "react";
+
+export type RecentSearchListProps = {
+  searchRef: RefObject<HTMLInputElement>;
+  setShowRecentSearch: (show: boolean) => void;
+};
